+++ assets/jsx/classic-editor.jsx
@@ -1,8 +1,8 @@
 import { FutureActionPanelClassicEditor } from './components';
 import { createStore } from './data';
 import { isGutenbergEnabled } from './utils';
-import { createRoot } from '@wordpress/element';
 import { select } from '@wordpress/data';
+import { createRoot } from 'react-dom/client';
 import {
     postType,
     isNewPost,
@@ -14,9 +14,9 @@ import {
     taxonomyName,
     postTypeDefaultConfig,
     defaultDate,
-    statusesSelectOptions
+    statusesSelectOptions,
+    hideCalendarByDefault
 } from "&config.classic-editor";
-import { render } from "react-dom";
 
 if (! isGutenbergEnabled()) {
     const storeName = 'publishpress-future/future-action';
@@ -48,6 +48,7 @@ if (! isGutenbergEnabled()) {
             startOfWeek={startOfWeek}
             strings={strings}
             taxonomyName={taxonomyName}
+            hideCalendarByDefault={hideCalendarByDefault}
         />
     );