+++ src/Views/menu-general.php
@@ -50,6 +50,8 @@ if ($expiredemailnotificationadmins == 0) {
     $expiredemailnotificationadminsenabled = 'checked="checked"';
 }
 
+$calendarHiddenByDefault = (bool) get_option('expirationdateHideCalendarByDefault', false);
+
 $user_roles = wp_roles()->get_names();
 $plugin_facade = PostExpirator_Facade::getInstance();
 ?>
@@ -83,7 +85,29 @@ $plugin_facade = PostExpirator_Facade::getInstance();
                                     '<code>',
                                     '</code>'
                                 ); ?></p>
+
+                                <div id="expiration-date-preview"></div>
                         </div>
+                    </td>
+                </tr>
+                <tr valign="top">
+                    <th scope="row"><label for="expired-hide-calendar-by-default"><?php
+                            esc_html_e('Calendar Visibility', 'post-expirator'); ?></label></th>
+                    <td>
+                        <div class="pp-settings-field-row">
+                            <input type="radio" name="expired-hide-calendar-by-default" id="expired-hide-calendar-by-default-false"
+                                value="0" <?php echo $calendarHiddenByDefault ? '' : 'checked="checked"'; ?>/> <label
+                                    for="expired-hide-calendar-by-default-false"><?php
+                                    esc_html_e('Remember last state', 'post-expirator'); ?></label>
+                        </div>
+                        <p class="description"><?php esc_html_e('Shows or hides the calendar based on the last user interaction.', 'post-expirator'); ?></p>
+                        <div class="pp-settings-field-row">
+                            <input type="radio" name="expired-hide-calendar-by-default" id="expired-hide-calendar-by-default-true"
+                                value="1" <?php echo $calendarHiddenByDefault ? 'checked="checked"' : ''; ?>/> <label
+                                    for="expired-hide-calendar-by-default-true"><?php
+                                    esc_html_e('Always hidden', 'post-expirator'); ?></label>
+                        </div>
+                        <p class="description"><?php esc_html_e('The calendar is always hidden by default.', 'post-expirator'); ?></p>
                     </td>
                 </tr>
             </table>